Computing distances from training point in parallel
The bottleneck in go-knn's performance has been how long it takes to do distance
computation. For each point that it tests/ trains on, it must compute the
distance to every other point in that training/ test set. Since we compute all
of these distances before we change any weights, the computations can be
parallelized, which is what this commit accomplishes via goroutines and
WaitGroups. In my tests on an Intel i7-5600U processor, this speeds up the total
computation time by just under 3x.